requirements.txt
aiofiles==0.5.0
aioredis==1.3.1
async-timeout==3.0.1
certifi==2020.4.5.1
chardet==3.0.4
contextvars==2.4
dill==0.3.1.1
futures==3.1.1
greenlet==0.4.15
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hiredis==1.0.1
hpack==3.0.0
hstspreload==2020.5.19
httptools==0.1.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.9
immutables==0.14
multidict==4.7.6
pip==20.0.2
pycryptodome==3.9.7
rfc3986==1.4.0
sanic==20.3.0
Sanic-Cors==0.10.0.post3
Sanic-Plugins-Framework==0.9.2
sanic-redis==0.2.0
setuptools==44.1.0
sniffio==1.1.0
ujson==2.0.3
urllib3==1.25.9
websockets==8.1
